2015-02-20 Roman - DAQ.Fastwell: Addition headers included to distribution. - FIX:DAQ.Siemens: Into Libnodave file nodave.h #pragma pack(1) is replaced to #pragma pack(push,1) ... #pragma pack(pop). - FIX:DAQ.Siemens,Special.SystemTests: Integer type expanded next to int64_t. - UI.{VCAEngine,Vision}: Some warnings added for try add already present items and paste, after copy or cut, to already present item.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2214 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/daq/Siemens/siemens.cpp
@@ -1221,7 +1221,7 @@ string TMdContr::getValS( SValData ival, ResString &err )
 
 void TMdContr::setValB( bool ivl, SValData ival, ResString &err )
 {
-    int val = getValI(SValData(ival.db,ival.off,1), err);
+    int64_t val = getValI(SValData(ival.db,ival.off,1), err);
     if(val == EVAL_INT || (bool)(val&(0x01<<ival.sz)) == ivl) return;
     //Write data to controller or write data block
     val ^= (0x01<<ival.sz);
@@ -1246,7 +1246,7 @@ void TMdContr::setValB( bool ivl, SValData ival, ResString &err )
 
 void TMdContr::setValI( int64_t ivl, SValData ival, ResString &err )
 {
-    int val = getValI(ival, err);
+    int64_t val = getValI(ival, err);
     if(val == EVAL_INT || val == ivl) return;
     //Write data to controller or write data block
     val = ivl;
